+++ YoutubeDownloader.Core/Downloading/ClosedCaptionsDownloader.cs
@@ -12,10 +12,18 @@ namespace YoutubeDownloader.Core.Downloading
     public class ClosedCaptionsDownloader
     {
         private readonly YoutubeClient _youtube = new(Http.Client);
-        public async Task<Tuple<bool, string>> DownloadCCAsync(string path, IVideo video, CancellationToken cancellationToken = default)
+
+        public async Task<Tuple<bool, string>> DownloadCCAsync(
+            string path,
+            IVideo video,
+            CancellationToken cancellationToken = default
+        )
         {
             var tempPath = Path.ChangeExtension(path, "srt");
-            var manifest = await _youtube.Videos.ClosedCaptions.GetManifestAsync(video.Id, cancellationToken);
+            var manifest = await _youtube.Videos.ClosedCaptions.GetManifestAsync(
+                video.Id,
+                cancellationToken
+            );
             bool isChineseCC = false;
             if (manifest.Tracks.Count != 0)
             {
@@ -29,7 +37,11 @@ namespace YoutubeDownloader.Core.Downloading
                 {
                     trackInfo = manifest.GetByLanguage("en");
                 }
-                await _youtube.Videos.ClosedCaptions.DownloadAsync(trackInfo, tempPath, cancellationToken: cancellationToken);
+                await _youtube.Videos.ClosedCaptions.DownloadAsync(
+                    trackInfo,
+                    tempPath,
+                    cancellationToken: cancellationToken
+                );
                 return Tuple.Create(isChineseCC, tempPath);
             }
             return Tuple.Create(isChineseCC, "");