+++ YoutubeDownloader.Core/Resolving/QueryResolver.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Net;
 using System.Threading;
 using System.Threading.Tasks;
 using Gress;
@@ -15,14 +16,18 @@ namespace YoutubeDownloader.Core.Resolving;
 
 public class QueryResolver
 {
-    private readonly YoutubeClient _youtube = new(Http.Client);
+    private readonly YoutubeClient _youtube;
+
+    public QueryResolver(IReadOnlyList<Cookie>? initialCookies = null) =>
+        _youtube = new YoutubeClient(Http.Client, initialCookies ?? Array.Empty<Cookie>());
 
     public async Task<QueryResult> ResolveAsync(
         string query,
-        CancellationToken cancellationToken = default)
+        CancellationToken cancellationToken = default
+    )
     {
-        // Only consider URLs when parsing IDs.
-        // All other queries should be treated as search queries.
+        // Only consider URLs for parsing IDs.
+        // All other queries should be treated as search keywords.
         var isUrl = Uri.IsWellFormedUriString(query, UriKind.Absolute);
 
         // Playlist
@@ -37,7 +42,7 @@ public class QueryResolver
         if (isUrl && VideoId.TryParse(query) is { } videoId)
         {
             var video = await _youtube.Videos.GetAsync(videoId, cancellationToken);
-            return new QueryResult(QueryResultKind.Video, video.Title, new[] {video});
+            return new QueryResult(QueryResultKind.Video, video.Title, new[] { video });
         }
 
         // Channel
@@ -51,7 +56,10 @@ public class QueryResolver
         // Channel (by handle)
         if (isUrl && ChannelHandle.TryParse(query) is { } channelHandle)
         {
-            var channel = await _youtube.Channels.GetByHandleAsync(channelHandle, cancellationToken);
+            var channel = await _youtube.Channels.GetByHandleAsync(
+                channelHandle,
+                cancellationToken
+            );
             var videos = await _youtube.Channels.GetUploadsAsync(channel.Id, cancellationToken);
             return new QueryResult(QueryResultKind.Channel, $"Channel: {channel.Title}", videos);
         }
@@ -74,7 +82,9 @@ public class QueryResolver
 
         // Search
         {
-            var videos = await _youtube.Search.GetVideosAsync(query, cancellationToken).CollectAsync(20);
+            var videos = await _youtube.Search
+                .GetVideosAsync(query, cancellationToken)
+                .CollectAsync(20);
             return new QueryResult(QueryResultKind.Search, $"Search: {query}", videos);
         }
     }
@@ -82,7 +92,8 @@ public class QueryResolver
     public async Task<QueryResult> ResolveAsync(
         IReadOnlyList<string> queries,
         IProgress<Percentage>? progress = null,
-        CancellationToken cancellationToken = default)
+        CancellationToken cancellationToken = default
+    )
     {
         if (queries.Count == 1)
             return await ResolveAsync(queries.Single(), cancellationToken);
@@ -91,7 +102,6 @@ public class QueryResolver
         var videoIds = new HashSet<VideoId>();
 
         var completed = 0;
-
         foreach (var query in queries)
         {
             var result = await ResolveAsync(query, cancellationToken);
@@ -102,9 +112,7 @@ public class QueryResolver
                     videos.Add(video);
             }
 
-            progress?.Report(
-                Percentage.FromFraction(1.0 * ++completed / queries.Count)
-            );
+            progress?.Report(Percentage.FromFraction(1.0 * ++completed / queries.Count));
         }
 
         return new QueryResult(QueryResultKind.Aggregate, $"{queries.Count} queries", videos);