+++ YoutubeDownloader.Core/Utils/Extensions/GenericExtensions.cs
@@ -2,7 +2,8 @@
 
 namespace YoutubeDownloader.Core.Utils.Extensions;
 
-internal static class GenericExtensions
+public static class GenericExtensions
 {
-    public static TOut Pipe<TIn, TOut>(this TIn input, Func<TIn, TOut> transform) => transform(input);
+    public static TOut Pipe<TIn, TOut>(this TIn input, Func<TIn, TOut> transform) =>
+        transform(input);
 }