+++ YoutubeDownloader.Core/Utils/Extensions/StringExtensions.cs
@@ -1,9 +1,7 @@
 ﻿namespace YoutubeDownloader.Core.Utils.Extensions;
 
-internal static class StringExtensions
+public static class StringExtensions
 {
     public static string? NullIfEmptyOrWhiteSpace(this string str) =>
-        !string.IsNullOrEmpty(str.Trim())
-            ? str
-            : null;
+        !string.IsNullOrEmpty(str.Trim()) ? str : null;
 }