+++ YoutubeDownloader.Core/Utils/Extensions/YoutubeExtensions.cs
@@ -3,7 +3,7 @@ using YoutubeExplode.Common;
 
 namespace YoutubeDownloader.Core.Utils.Extensions;
 
-internal static class YoutubeExtensions
+public static class YoutubeExtensions
 {
     public static string? TryGetImageFormat(this Thumbnail thumbnail) =>
         Url.TryExtractFileName(thumbnail.Url)?.Pipe(Path.GetExtension)?.Trim('.');