+++ YoutubeDownloader.Core/Utils/Http.cs
@@ -4,20 +4,21 @@ using System.Reflection;
 
 namespace YoutubeDownloader.Core.Utils;
 
-internal static class Http
+public static class Http
 {
-    public static HttpClient Client { get; } = new()
-    {
-        DefaultRequestHeaders =
+    public static HttpClient Client { get; } =
+        new()
         {
-            // Required by some of the services we're using
-            UserAgent =
+            DefaultRequestHeaders =
             {
-                new ProductInfoHeaderValue(
-                    "YoutubeDownloader",
-                    Assembly.GetExecutingAssembly().GetName().Version?.ToString(3)
-                )
+                // Required by some of the services we're using
+                UserAgent =
+                {
+                    new ProductInfoHeaderValue(
+                        "YoutubeDownloader",
+                        Assembly.GetExecutingAssembly().GetName().Version?.ToString(3)
+                    )
+                }
             }
-        }
-    };
+        };
 }