+++ YoutubeDownloader/Converters/InverseBoolConverter.cs
@@ -12,6 +12,10 @@ public class InverseBoolConverter : IValueConverter
     public object Convert(object? value, Type targetType, object parameter, CultureInfo culture) =>
         value is false;
 
-    public object ConvertBack(object? value, Type targetType, object parameter, CultureInfo culture) =>
-        value is false;
+    public object ConvertBack(
+        object? value,
+        Type targetType,
+        object parameter,
+        CultureInfo culture
+    ) => value is false;
 }