+++ YoutubeDownloader/Converters/VideoToLowestQualityThumbnailUrlConverter.cs
@@ -12,10 +12,12 @@ public class VideoToLowestQualityThumbnailUrlConverter : IValueConverter
     public static VideoToLowestQualityThumbnailUrlConverter Instance { get; } = new();
 
     public object? Convert(object? value, Type targetType, object parameter, CultureInfo culture) =>
-        value is IVideo video
-            ? video.Thumbnails.OrderBy(t => t.Resolution.Area).FirstOrDefault()?.Url
-            : null;
+        value is IVideo video ? video.Thumbnails.MinBy(t => t.Resolution.Area)?.Url : null;
 
-    public object ConvertBack(object? value, Type targetType, object parameter, CultureInfo culture) =>
-        throw new NotSupportedException();
+    public object ConvertBack(
+        object? value,
+        Type targetType,
+        object parameter,
+        CultureInfo culture
+    ) => throw new NotSupportedException();
 }