+++ YoutubeDownloader/Utils/ProcessEx.cs
@@ -7,10 +7,7 @@ internal static class ProcessEx
 {
     public static void Start(string path, IReadOnlyList<string>? arguments = null)
     {
-        using var process = new Process
-        {
-            StartInfo = new ProcessStartInfo(path)
-        };
+        using var process = new Process { StartInfo = new ProcessStartInfo(path) };
 
         if (arguments is not null)
         {
@@ -25,10 +22,7 @@ internal static class ProcessEx
     {
         using var process = new Process
         {
-            StartInfo = new ProcessStartInfo(path)
-            {
-                UseShellExecute = true
-            }
+            StartInfo = new ProcessStartInfo(path) { UseShellExecute = true }
         };
 
         if (arguments is not null)