+++ YoutubeDownloader/Views/Dialogs/SettingsView.xaml.cs
@@ -9,9 +9,9 @@ public partial class SettingsView
         InitializeComponent();
     }
 
-    private void DarkModeToggleButton_Checked(object sender, RoutedEventArgs args) =>
+    private void DarkModeToggleButton_OnChecked(object sender, RoutedEventArgs args) =>
         App.SetDarkTheme();
 
-    private void DarkModeToggleButton_Unchecked(object sender, RoutedEventArgs args) =>
+    private void DarkModeToggleButton_OnUnchecked(object sender, RoutedEventArgs args) =>
         App.SetLightTheme();
 }