+++ Changelog.md
@@ -1,5 +1,36 @@
 # Changelog
 
+## v1.10.4 (21-Oct-2023)
+
+- Fixed various YouTube-related issues. Updated to YoutubeExplode v6.3.6.
+
+## v1.10.3 (31-Aug-2023)
+
+- Fixed various YouTube-related issues. Updated to YoutubeExplode v6.3.3.
+
+## v1.10.2 (18-Aug-2023)
+
+- Fixed various YouTube-related issues. Updated to YoutubeExplode v6.3.2.
+
+## v1.10.1 (15-Jul-2023)
+
+- Fixed an issue which prevented authentication from working properly.
+
+## v1.10 (10-Jul-2023)
+
+- Added support for fetching private videos and playlists. You can click the "Authentication" button in the top right corner to log in with your YouTube account, which will allow you to access private or restricted content. If you don't wish your authentication data (cookies) to be persisted on disk, toggle the "Persist authentication" option off in settings. (Thanks [@xBaank](https://github.com/xBaank))
+- Fixed various YouTube-related issues. Updated to YoutubeExplode v6.2.17.
+
+## v1.9.13 (28-Jun-2023)
+
+- Changed how failed downloads are presented in the UI. The error message corresponding to a failed download is no longer shown in the tooltip when you hover over the "Failed" status text. Instead, you can click on the status text to copy the error message to the clipboard. This should make it easier to include the error in a bug report.
+- Fixed an issue which prevented videos longer than 24 hours from being downloaded.
+- Fixed various YouTube-related issues. Updated to YoutubeExplode v6.2.16.
+
+## v1.9.12 (25-May-2023)
+
+- Fixed various YouTube-related issues. Updated to YoutubeExplode v6.2.15.
+
 ## v1.9.11 (02-May-2023)
 
 - Fixed various YouTube-related issues. Updated to YoutubeExplode v6.2.14.